Added testing for part 5
