+++ templates/edit_movies_form.html
@@ -1,7 +1,6 @@
 {% extends '_layout.html' %}
 
 {% block title %}Movie Rating App | Edit Movie{% endblock %}
-
 {% block body %}
 <h1 class="mb-5">Edit Movie</h1>
 <form method="POST" action="{{ url_for('update_movie', movie_id=movie.movie_id) }}">
@@ -12,7 +11,6 @@
     <input type="text" class="form-control" id="director" name="director" value="{{ movie.director }}" required>
     <br>
     <label for="rating">Rating</label> 
-
     <select class="form-control" id="rating" name="rating" required>
       <option value="1" {% if movie.rating == 1 %}selected{% endif %}>1</option>
       <option value="2" {% if movie.rating == 2 %}selected{% endif %}>2</option>
@@ -20,9 +18,7 @@
       <option value="4" {% if movie.rating == 4 %}selected{% endif %}>4</option>
       <option value="5" {% if movie.rating == 5 %}selected{% endif %}>5</option>
     </select>
-    
     <br>
-
     <button type="submit" class="btn btn-primary">Save</button>
   </form>  
   {% endblock %}
